refactor(feature): refactor daily-log module name - daily-log의 모듈명을 history로 변경하고 전부 수정합니다.
