+++ core/designsystem/src/main/kotlin/com/gradation/lift/designsystem/resource/Icon.kt
@@ -5,7 +5,7 @@ import com.gradation.lift.designsystem.R
 
 object LiftIcon {
     val Home = R.drawable.home
-    val DailyLog = R.drawable.history
+    val History = R.drawable.history
     val Analytics = R.drawable.analytics
     val MyInfo = R.drawable.my_info
     val CheckBoxChecked = R.drawable.checkbox_checked
