+++ core/navigation/src/main/kotlin/com/gradation/lift/navigation/Route.kt
@@ -8,7 +8,7 @@ package com.gradation.lift.navigation
  * [BADGE_GRAPH_NAME]
  * [CREATE_ROUTINE_GRAPH_NAME]
  * [UPDATE_ROUTINE_GRAPH_NAME]
- * [DAILY_LOG_GRAPH_NAME]
+ * [HISTORY_GRAPH_NAME]
  * [ANALYTICS_GRAPH_NAME]
  * [HOME_GRAPH_NAME]
  * [LOGIN_GRAPH_NAME]
@@ -44,8 +44,8 @@ object Route {
     const val UPDATE_ROUTINE_FIND_WORK_CATEGORY_ROUTER_NAME = "update_routine_find_work_category"
     const val UPDATE_ROUTINE_CHANGE_ORDER_ROUTER_NAME = "update_routine_change_order"
 
-    const val DAILY_LOG_GRAPH_NAME = "daily_log_graph"
-    const val DAILY_LOG_DAILY_LOG_ROUTER_NAME = "daily_log_daily_log"
+    const val HISTORY_GRAPH_NAME = "history_graph"
+    const val HISTORY_HISTORY_ROUTER_NAME = "history_history"
 
     const val ANALYTICS_GRAPH_NAME = "analytics_graph"
     const val ANALYTICS_ANALYTICS_ROUTER_NAME = "analytics_analytics"
