+++ core/navigation/src/main/kotlin/com/gradation/lift/navigation/navigation/HomeNavigation.kt
@@ -4,7 +4,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import com.gradation.lift.navigation.Route
 import com.gradation.lift.navigation.Route.ANALYTICS_GRAPH_NAME
-import com.gradation.lift.navigation.Route.DAILY_LOG_GRAPH_NAME
+import com.gradation.lift.navigation.Route.HISTORY_GRAPH_NAME
 import com.gradation.lift.navigation.Route.HOME_GRAPH_NAME
 import com.gradation.lift.navigation.Route.ROUTINE_DETAIL_ROUTINE_ROUTER_NAME
 
@@ -73,8 +73,8 @@ fun NavController.navigateMyInfoGraph() {
     }
 }
 
-fun NavHostController.navigateDailyLogGraph() {
-    this.navigate(DAILY_LOG_GRAPH_NAME) {
+fun NavHostController.navigateHistoryGraph() {
+    this.navigate(HISTORY_GRAPH_NAME) {
         launchSingleTop = true
         popUpTo(currentDestination!!.id) {
             saveState = true
