--- feature/daily-log/daily-log/build.gradle.kts
@@ -1,22 +1,0 @@
-plugins {
-    id("lift.android.compose.library")
-    id("lift.android.library")
-    id("lift.android.kotlin")
-    id("lift.android.hilt")
-}
-
-android {
-    namespace = "com.gradation.lift.feature.dailyLog.dailyLog"
-}
-
-
-
-dependencies {
-    implementation(project(":core:domain"))
-    implementation(project(":core:model"))
-    implementation(project(":core:common"))
-    implementation(project(":core:designsystem"))
-    implementation(project(":core:test"))
-    implementation(project(":core:navigation"))
-    implementation(project(":core:ui"))
-}
