--- feature/daily-log/daily-log/src/main/kotlin/com/gradation/lift/feature/dailyLog/dailyLog/data/state/HistoryUiState.kt
@@ -1,11 +1,0 @@
-package com.gradation.lift.feature.dailyLog.dailyLog.data.state
-
-import com.gradation.lift.model.model.history.History
-
-sealed interface HistoryUiState {
-
-    data class Success(val historyList: List<History>) : HistoryUiState
-    data class Fail(val message: String) : HistoryUiState
-    data object Loading : HistoryUiState
-    data object Empty : HistoryUiState
-}
