--- feature/daily-log/daily-log/src/main/kotlin/com/gradation/lift/feature/dailyLog/dailyLog/ui/component/EmptyHistoryView.kt
@@ -1,8 +1,0 @@
-package com.gradation.lift.feature.dailyLog.dailyLog.ui.component
-
-import androidx.compose.runtime.Composable
-
-@Composable
-internal fun EmptyHistoryView(){
-
-}
