--- feature/daily-log/navigation/build.gradle.kts
@@ -1,16 +1,0 @@
-plugins {
-    id("lift.android.compose.library")
-    id("lift.android.library")
-    id("lift.android.kotlin")
-}
-
-android {
-    namespace = "com.gradation.lift.feature.dailyLog.navigation"
-}
-
-dependencies {
-    implementation(project(":core:designsystem"))
-    implementation(project(":core:navigation"))
-
-    implementation(project(":feature:daily-log:daily-log"))
-}
