+++ feature/history/history/src/main/AndroidManifest.xml
