+++ feature/history/history/src/main/kotlin/com/gradation/lift/feature/history/history/data/state/HistoryUiState.kt
@@ -1,0 +1,11 @@
+package com.gradation.lift.feature.history.history.data.state
+
+import com.gradation.lift.model.model.history.History
+
+sealed interface HistoryUiState {
+
+    data class Success(val historyList: List<History>) : HistoryUiState
+    data class Fail(val message: String) : HistoryUiState
+    data object Loading : HistoryUiState
+    data object Empty : HistoryUiState
+}
