+++ feature/history/history/src/main/kotlin/com/gradation/lift/feature/history/history/ui/component/EmptyHistoryView.kt
@@ -1,0 +1,8 @@
+package com.gradation.lift.feature.history.history.ui.component
+
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun EmptyHistoryView(){
+
+}
