+++ feature/history/navigation/build.gradle.kts
@@ -1,0 +1,15 @@
+plugins {
+    id("lift.android.compose.library")
+    id("lift.android.library")
+    id("lift.android.kotlin")
+}
+
+android {
+    namespace = "com.gradation.lift.feature.history.navigation"
+}
+
+dependencies {
+    implementation(project(":core:designsystem"))
+    implementation(project(":core:navigation"))
+    implementation(project(":feature:history:history"))
+}
