+++ feature/history/navigation/src/main/AndroidManifest.xml
