+++ feature/history/navigation/src/main/kotlin/com/gradation/lift/feature/history/navigation/HistoryGraphBuilder.kt
@@ -1,0 +1,31 @@
+package com.gradation.lift.feature.history.navigation
+
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.navigation
+import com.gradation.lift.feature.history.history.navigation.historyScreen
+import com.gradation.lift.navigation.Route.HISTORY_HISTORY_ROUTER_NAME
+import com.gradation.lift.navigation.Route.HISTORY_GRAPH_NAME
+
+
+fun NavGraphBuilder.historyGraphBuilder(
+    modifier: Modifier=Modifier,
+    navController: NavController,
+) {
+    navigation(
+        route = HISTORY_GRAPH_NAME,
+        startDestination = HISTORY_HISTORY_ROUTER_NAME,
+        popEnterTransition = {fadeIn()},
+        popExitTransition = { fadeOut()},
+        enterTransition = {fadeIn()},
+        exitTransition = { fadeOut()},
+    ) {
+        historyScreen(modifier)
+    }
+}
+
+
+
